# crates/trapfalld/src/openapi.yaml
openapi: 3.0.3
info:
  title: TrapFall API
  description: |
    Lightweight self-hosted error capture engine. Sentry SDK compatible.

    ## Authentication
    Dashboard API uses session cookies (`trapfall_session`). Obtain via `POST /api/0/auth/login`.

    Ingest API uses DSN key auth via `X-Sentry-Auth` header or `Authorization: Bearer <key>`.
  version: 0.0.3
  contact:
    name: TrapFall
    url: https://github.com/codecoradev/trapfall
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/0/ws:
    get:
      tags: [System]
      summary: WebSocket real-time updates
      description: |
        Upgrade to WebSocket. Requires session cookie auth.
        Sends `IssueCreated` and `IssueUpdated` messages as JSON.
      operationId: websocket
      security:
        - cookieAuth: []
      responses:
        "101":
          description: Switching Protocols (WebSocket upgrade)
        "401":
          description: Not authenticated
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: trapfall_session
    dsnKey:
      type: apiKey
      in: header
      name: X-Sentry-Auth
...